rename package, add launch.json for debugging, import internal package correctly
